# Conflicts:
#	server/src/main/java/de/ingrid/igeserver/profiles/ingrid/exporter/IngridModelTransformer.kt

+++ server/src/main/java/de/ingrid/igeserver/profiles/ingrid/importer/GeoserviceMapper.kt
@@ -1,3 +1,22 @@
+/**
+ * ==================================================
+ * Copyright (C) 2023-2024 wemove digital solutions GmbH
+ * ==================================================
+ * Licensed under the EUPL, Version 1.2 or – as soon they will be
+ * approved by the European Commission - subsequent versions of the
+ * EUPL (the "Licence");
+ *
+ * You may not use this work except in compliance with the Licence.
+ * You may obtain a copy of the Licence at:
+ *
+ * https://joinup.ec.europa.eu/software/page/eupl
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the Licence is distributed on an "AS IS" basis,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the Licence for the specific language governing permissions and
+ * limitations under the Licence.
+ */
 package de.ingrid.igeserver.profiles.ingrid.importer
 
 import de.ingrid.igeserver.exports.iso.Metadata
@@ -134,10 +153,10 @@ class GeoserviceMapper(metadata: Metadata, codeListService: CodelistHandler, cat
 
         return (0 until biggestListSize).map {
             Resolution(
-                scale.getOrNull(it)?.split(":")?.getOrNull(1)?.trim()?.toInt(), // "1:1000"
+                scale.getOrNull(it)?.split(":")?.getOrNull(1)?.trim()?.toFloat(), // "1:1000"
                 groundResolution.getOrNull(it)?.substring(0, groundResolution.getOrNull(it)?.length?.minus(1)!!)?.trim()
-                    ?.toInt(),
-                scanResolution.getOrNull(it)?.trim()?.toInt()
+                    ?.toFloat(),
+                scanResolution.getOrNull(it)?.trim()?.toFloat()
             )
         }
     }
